requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0

# Data augmentation
albumentations>=1.3.0
scipy>=1.10.0

# Training utilities
tqdm>=4.65.0
PyYAML>=6.0
tensorboard>=2.12.0

# Evaluation metrics
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Optional: Advanced SSM support
einops>=0.6.0

# Optional: Experiment tracking
wandb>=0.15.0
